So, it doesn't work after all...
In order to make it work, I think there needs to be a better way of creating the assembly. I should look over to either see how IL2CPPDumper does it, or just read it in as raw information and modify+write out. Clearly, there are some issues with writing to the assemblies they come from, although it is a little strange that there is... Maybe I need to look into this in more detail later, because it should be okay to just write to the existing IL2CPPDumper assemblies, but for whatever reason, I was getting string mismatch exceptions (probably due to the fact that my ReadAlignedString and fields used for strings are slightly different) not sure why though.