Made sure to import literally everything, also imports method calls
First, this might not actually fix the whole string problem (even though it goes away) because there is probably something wrong (like two referenced mscorlibs or something) and that would cause even more problems down the line (when trying to read the assembly). It seems to occur when there is an array of type string made.
Second, when double checking to see if dnspy created everything correctly, I could not actually compile the created code into c# (from the IL2CPPDumper library, it worked fine with the Unity one) which was due to the fact that there were 2 Action<T1,T2> objects, one in mscorlib 3.7.1.6, and one in mscorlib 4.0.0.0. I think the 4.0.0.0 comes from the fact that IL2CPP Dumper DLLs are created from .NET 4.0.0.0, so when the string is imported, it is importing a 4.0.0.0 string.
Third, I am getting the whole "ECall methods must be packaged into a system module" when attempting to construct the classes that I have created (even without the broken string array parsing [it needs and should use actions instead of defaulting to ReadFrom using reflection] and whatnot) I was thinking that I could delete all methods and see what happens, but now I'm not too sure. I think there might be something happening behind the scenes as to where I can see in UnityEngine which might require me to completely remove it when generating new asset types and make my own custom, smaller, safer, unityengine (not sure yet though, methods are the easiest to test and the highest possible reward)